fix(no_expect_outside_tests): restore rstest companion module detection after rebase
Move rstest companion test function detection from the
no_expect_outside_tests driver into the shared whitaker::hir module,
fixing a scoping bug: the original used hir_crate_items().free_items()
which flattens items across all modules, allowing false sibling matches
between items in different modules. The shared version starts from
hir_root_module().item_ids and recurses per-module.
