+++ crates/no_expect_outside_tests/src/driver/mod.rs
@@ -17,7 +17,7 @@ use log::debug;
 use rustc_hir as hir;
 use rustc_lint::{LateContext, LateLintPass};
 use rustc_middle::ty::{self, Ty};
-use rustc_span::{Span, Symbol, sym};
+use rustc_span::sym;
 use serde::Deserialize;
 use whitaker::SharedConfig;
 use whitaker::hir::has_test_like_hir_attributes;
@@ -69,7 +69,7 @@ impl<'tcx> LateLintPass<'tcx> for NoExpectOutsideTests {
         self.is_test_harness = cx.tcx.sess.opts.test;
         self.harness_marked_test_functions = if self.is_test_harness {
             let mut marked = whitaker::hir::collect_harness_test_functions(cx);
-            collect_rstest_companion_test_functions(cx, &mut marked);
+            marked.extend(whitaker::hir::collect_rstest_companion_test_functions(cx));
             marked
         } else {
             HashSet::new()
@@ -247,128 +247,6 @@ fn is_harness_marked_test_function(
     harness_marked_test_functions: &HashSet<hir::HirId>,
 ) -> bool {
     harness_marked_test_functions.contains(&function_hir_id)
-}
-
-/// Augments a harness-marked set with functions whose rstest companion module
-/// contains a harness descriptor.
-///
-/// The shared `whitaker::hir::collect_harness_test_functions` already catches
-/// direct const-descriptor siblings. This helper additionally finds functions
-/// with a same-named sibling *module* that itself contains a harness descriptor
-/// (the pattern rstest case expansions produce).
-fn collect_rstest_companion_test_functions<'tcx>(
-    cx: &LateContext<'tcx>,
-    marked: &mut HashSet<hir::HirId>,
-) {
-    let root_items: Vec<_> = cx
-        .tcx
-        .hir_crate_items(())
-        .free_items()
-        .map(|id| cx.tcx.hir_item(id))
-        .collect();
-    collect_companion_in_group(cx, &root_items, marked);
-}
-
-fn collect_companion_in_group<'tcx>(
-    cx: &LateContext<'tcx>,
-    items: &[&'tcx hir::Item<'tcx>],
-    marked: &mut HashSet<hir::HirId>,
-) {
-    for item in items
-        .iter()
-        .copied()
-        .filter(|item| matches!(item.kind, hir::ItemKind::Fn { .. }))
-    {
-        let Some(ident) = item.kind.ident() else {
-            continue;
-        };
-
-        if items
-            .iter()
-            .copied()
-            .any(|sibling| has_companion_test_module(cx, item.hir_id(), ident.name, sibling))
-        {
-            marked.insert(item.hir_id());
-        }
-    }
-
-    for item in items {
-        let hir::ItemKind::Mod(_, module) = item.kind else {
-            continue;
-        };
-
-        let module_items: Vec<_> = module
-            .item_ids
-            .iter()
-            .map(|id| cx.tcx.hir_item(*id))
-            .collect();
-        collect_companion_in_group(cx, &module_items, marked);
-    }
-}
-
-fn has_companion_test_module<'tcx>(
-    cx: &LateContext<'tcx>,
-    function_hir_id: hir::HirId,
-    function_name: Symbol,
-    sibling: &'tcx hir::Item<'tcx>,
-) -> bool {
-    sibling.hir_id() != function_hir_id
-        && sibling
-            .kind
-            .ident()
-            .is_some_and(|ident| ident.name == function_name)
-        && matches!(sibling.kind, hir::ItemKind::Mod(..))
-        && module_has_harness_descriptor(cx, sibling)
-}
-
-fn module_has_harness_descriptor<'tcx>(
-    cx: &LateContext<'tcx>,
-    module_item: &'tcx hir::Item<'tcx>,
-) -> bool {
-    let hir::ItemKind::Mod(_, module) = module_item.kind else {
-        return false;
-    };
-
-    let items: Vec<_> = module
-        .item_ids
-        .iter()
-        .map(|item_id| cx.tcx.hir_item(*item_id))
-        .collect();
-
-    items
-        .iter()
-        .copied()
-        .filter(|item| matches!(item.kind, hir::ItemKind::Fn { .. }))
-        .any(|fn_item| {
-            let Some(fn_ident) = fn_item.kind.ident() else {
-                return false;
-            };
-
-            let fn_hir_id = fn_item.hir_id();
-            let fn_name = fn_ident.name;
-            let fn_span = fn_item.span;
-            items
-                .iter()
-                .copied()
-                .any(|sibling| is_harness_const_descriptor(fn_hir_id, fn_name, fn_span, sibling))
-        })
-}
-
-fn is_harness_const_descriptor(
-    function_hir_id: hir::HirId,
-    function_name: Symbol,
-    function_span: Span,
-    sibling: &hir::Item<'_>,
-) -> bool {
-    // `rustc --test` may synthesize a const descriptor that shares the test
-    // function's name and source range. The wrapper function and descriptor can
-    // carry different syntax contexts, so this must compare source bytes
-    // rather than exact `Span` identity.
-    sibling.hir_id() != function_hir_id
-        && matches!(sibling.kind, hir::ItemKind::Const(..))
-        && sibling.kind.ident().is_some_and(|ident| {
-            ident.name == function_name && sibling.span.source_equal(function_span)
-        })
 }
 
 // Check if any attribute is #[test].
